# UserService/public/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "title": "User Service API.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/{username}": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Получение информации о пользователе.",
        "operationId": "GetUserInfoRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/UserInfoResponse"
          },
          "400": {
            "$ref": "#/responses/StatusBadRequest"
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Обновление информации о пользователе.",
        "operationId": "UpdateUserRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "user",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserForUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/updateuserResponse"
          },
          "400": {
            "$ref": "#/responses/StatusBadRequest"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "user"
        ],
        "summary": "Удаление пользователя.",
        "operationId": "DeleteUserRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/DeleteUserResponse"
          },
          "400": {
            "$ref": "#/responses/StatusBadRequest"
          }
        }
      },
      "patch": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "admin"
        ],
        "summary": "Admin: Изменение баланса пользователя.",
        "operationId": "ChangeBalanceRequest",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/BalanceChange"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ChangeBalanceResponse"
          },
          "400": {
            "$ref": "#/responses/StatusBadRequest"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BalanceChange": {
      "type": "object",
      "properties": {
        "money": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Money"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "userService/internal/model"
    },
    "ResponseMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-package": "userService/internal/model"
    },
...
    "UserForUpdate": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/UserInfoForUpdate"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "userService/internal/model"
    },
    "UserInfo": {
      "type": "object",
      "properties": {
        "balance": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Balance"
        },
        "firstname": {
          "type": "string",
          "x-go-name": "Firstname"
        },
        "lastname": {
          "type": "string",
          "x-go-name": "Lastname"
        },
        "surname": {
          "type": "string",
          "x-go-name": "Surname"
        },
        "user_group": {
          "type": "string",
          "x-go-name": "Group"
        }
      },
      "x-go-package": "userService/internal/model"
    },
    "UserInfoForUpdate": {
      "type": "object",
      "properties": {
        "firstname": {
          "type": "string",
          "x-go-name": "Firstname"
        },
        "group": {
          "type": "string",
          "x-go-name": "Group"
        },
        "lastname": {
          "type": "string",
          "x-go-name": "Lastname"
        },
        "surname": {
          "type": "string",
          "x-go-name": "Surname"
        }
      },
      "x-go-package": "userService/internal/model"
    },
...
  },
  "responses": {
    "ChangeBalanceResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ResponseMessage"
      }
    },
...
    "DeleteUserResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ResponseMessage"
      }
    },
...
    "StatusBadRequest": {
      "description": "Bad Request replys an error of API calling."
    },
...
    "UserInfoResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/UserInfo"
      }
    },
...
    "updateuserResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ResponseMessage"
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}